new hook to control component
